+++ tests/Git.Repository.Tests/GitTests.cs
@@ -101,7 +101,7 @@ public class GitTests
 
         // Let's assume v2 pack index files
         Bucket index = new byte[] { 255, (byte)'t', (byte)'O', (byte)'c' }.AsBucket();
-        index += NetBitConverter.GetBytes((int)2).AsBucket();
+        index += NetBitConverter.GetBytes(2).AsBucket();
 
         // Fanout table
         index += fanOut.SelectMany(x => NetBitConverter.GetBytes(x)).ToArray().AsBucket();
